+++ lib/management/push.js
@@ -9,19 +9,14 @@ const assert = require('assert');
 const _config = require('../Config').config;
 const logger = require('../utilities/logger');
 const metadata = require('../metadata/wrapper');
+
 const { reshapeExceptionError } = arsenal.errorUtils;
-
-const {
-    patchConfiguration,
-    saveConfigurationVersion,
-} = require('./configuration');
-
+const { isManagementAgentUsed } = require('./agentClient');
+const { applyAndSaveOverlay } = require('./configuration');
 const {
     ChannelMessageV0,
     MessageType,
 } = require('./ChannelMessageV0');
-
-const { isManagementAgentUsed } = require('./agentClient');
 
 const {
     CONFIG_OVERLAY_MESSAGE,
@@ -185,28 +180,6 @@ function startWSManagementClient(pushEndpoint, instanceId, token) {
             socketsByChannelId[channelId] = socket;
         }
         socket.write(payload);
-    }
-
-    function applyAndSaveOverlay(overlay, log) {
-        patchConfiguration(overlay, log, err => {
-            if (err) {
-                log.error('could not apply pushed overlay', {
-                    error: reshapeExceptionError(err),
-                    method: 'applyAndSaveOverlay',
-                });
-                return;
-            }
-            saveConfigurationVersion(null, overlay, log, err => {
-                if (err) {
-                    log.error('could not cache overlay version', {
-                        error: reshapeExceptionError(err),
-                        method: 'applyAndSaveOverlay',
-                    });
-                    return;
-                }
-                log.info('overlay push processed');
-            });
-        });
     }
 
     function browserAccessChangeHandler() {
